requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.12.3
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
dnspython==2.7.0
feedparser==6.0.11
frozenlist==1.5.0
gnews==0.4.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai>=1.0.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
instagrapi==2.1.3
multidict==6.1.0
packaging==24.2
pillow==11.1.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
pyparsing==3.2.1
PySocks==1.7.1
python-dotenv==1.0.1
replicate==1.0.4
requests==2.32.3
rsa==4.9
setuptools==75.6.0
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wheel==0.45.1
yarl==1.18.3
streamlit>=1.32.0
streamlit-autorefresh>=1.0.1
